tests/wpt/web-platform-tests/origin-isolation/insecure-http.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Parent requests isolation, child requests isolation, child is a subdomain of the parent, but all over insecure HTTP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script type="module">
import { insertIframe, testSameAgentCluster } from "./resources/helpers.mjs";

promise_setup(async () => {
  await insertIframe("{{hosts[][www]}}", "?1");
});

// All isolation requests are ignored, since this is over insecure HTTP.
// So both end up in the site-keyed agent cluster.
testSameAgentCluster([self, 0]);

// Has to be promise_test because we used promise_setup().
promise_test(async () => {
  assert_false("originIsolationRestricted" in window);
}, "The getter must not exist");
</script>
